requirements.txt
imageio
imageio[ffmpeg]
mysql-connector-python
opencv-python-headless
openpyxl
pandas==1.5.3
Pillow==9.5.0
skia-python
SQLAlchemy
streamlit
streamlit-webrtc
twilio~=8.10.1
yolov5
